Notice
- I have to change my course in order to understand these.
What means?

LeetCode is good for programmers, harness their skills, however, in my case, It don't work for me because it requires for me to know about previous conception that need to solution the problem that I encounter this time.

So, It allows me to improve my skills at this point, I need to find alternative.