+
This need to be more example.
The example of this problem is so vague for solver to understand the logic of this. It is terrible.

For instance, the variable 'p' have to be  only the initial condition, but the answer require to be the place every recursive times. It is wrong. It have to be only one situation.